requirements.txt
torchvision>=0.8.1
ultralytics>=8.0.105
fastapi>=0.110.2
pillow>=10.3.0
pydantic>=2.7.1
matplotlib>=3.2.2
websockets>=12.0
opencv-python-headless